split multi select and select
