+++ projects/ui-framework/src/lib/form-elements/form-elements.module.ts
@@ -6,10 +6,10 @@ import { InputComponent } from './input/input.component';
 import { TextareaModule } from './textarea/textarea.module';
 import { TextareaComponent } from './textarea/textarea.component';
 import { SearchModule } from '../navigation/search/search.module';
-import { SelectModule } from './select-elements';
+import { SelectModule } from './select';
 import { CheckboxModule } from './checkbox/checkbox.module';
-import { MultiSelectComponent } from './select-elements/multi-select/multi-select.component';
-import { SingleSelectComponent } from './select-elements/single-select/single-select.component';
+import { MultiSelectComponent } from './select/multi-select/multi-select.component';
+import { SingleSelectComponent } from './select/single-select/single-select.component';
 import { CheckboxComponent } from './checkbox/checkbox.component';
 
 @NgModule({
