+++ projects/ui-framework/src/lib/form-elements/index.ts
@@ -1,6 +1,6 @@
 export * from './datepicker';
 export * from './input';
 export * from './textarea';
-export * from './select-elements';
+export * from './select';
 export * from './checkbox';
 export * from '../navigation/search';
