+++ projects/ui-framework/src/lib/form-elements/select/multi-select/multi-select.component.ts
@@ -1,42 +1,37 @@
 import { Component, EventEmitter, Input, OnInit, Output, ViewChild } from '@angular/core';
-import { concat, flatMap, get, join, escapeRegExp, forEach } from 'lodash';
+import { concat, escapeRegExp, flatMap, forEach, get, join } from 'lodash';
 import { SelectGroupOption, SelectionGroupOption, SelectionOption } from '../select.interface';
 import { ButtonSize, ButtonType } from '../../../buttons-indicators/buttons/buttons.enum';
-import { SelectModelService } from './select-model.service';
+import { SelectModelService } from '../select-model-service/select-model.service';
 import { BaseInputElement } from '../../base-input-element';
 import { IconColor, Icons, IconSize } from '../../../icons';
 
 const navigationKeys = new Set(['ArrowUp', 'ArrowDown', 'Enter']);
 
 @Component({
-  selector: 'b-select',
-  templateUrl: './select.component.html',
-  styleUrls: ['./select.component.scss'],
+  selector: 'b-multi-select',
+  templateUrl: './multi-select.component.html',
+  styleUrls: ['../select.scss'],
 })
-export class SelectComponent extends BaseInputElement implements OnInit {
+export class MultiSelectComponent extends BaseInputElement implements OnInit {
 
   @ViewChild('mySelect') mySelect;
   @ViewChild('triggerValueText') triggerValueText;
 
   @Input() options: SelectGroupOption[];
   @Input() value: (string | number)[] = [];
-  @Input() isMultiSelect: boolean;
   @Input() showSingleGroupHeader = false;
-  @Input() label: string;
-  @Input() required: boolean;
-  @Input() disabled: boolean;
-  @Input() hintMessage: string;
-  @Input() errorMessage: string;
 
   @Output() selectChange: EventEmitter<(string | number)[]> = new EventEmitter<(string | number)[]>();
 
   selectionGroupOptions: SelectionGroupOption[];
-  selectedModel: SelectionOption[] | SelectionOption;
+  selectedModel: SelectionOption[];
   triggerValue: string;
   blockGroupHeaderOptionClick = false;
   blockSelectClick = false;
   hasEllipsis = false;
-  panelClass: string;
+  panelClass = 'multi-select-panel';
+  isMultiSelect = true;
 
   readonly buttonType = ButtonType;
   readonly buttonSize = ButtonSize;
@@ -52,9 +47,6 @@ export class SelectComponent extends BaseInputElement implements OnInit {
   }
 
   ngOnInit(): void {
-    this.panelClass = this.isMultiSelect
-      ? 'multi-select-panel'
-      : 'single-select-panel';
     this.selectionGroupOptions = this.selectModelService
       .getSelectElementOptionsModel(this.options);
 
@@ -63,18 +55,17 @@ export class SelectComponent extends BaseInputElement implements OnInit {
     const selectedGroupsHeaders = this.selectModelService
       .getSelectedGroupHeaderOptions(this.selectionGroupOptions, this.value);
 
-    this.selectedModel = this.isMultiSelect
-      ? concat(selectedOptions, selectedGroupsHeaders)
-      : selectedOptions[0];
+    this.selectedModel = concat(selectedOptions, selectedGroupsHeaders);
 
     this.updateTriggerText();
   }
 
   onOptionClick(selectedOption): void {
-    if (this.isMultiSelect) {
-      this.onMultiSelectOptionClick(selectedOption);
+    if (selectedOption.isGroupHeader) {
+      this.onGroupHeaderClick(selectedOption);
     } else {
-      this.onSingleSelectOptionClick(selectedOption);
+      this.selectedModel = this.selectModelService
+        .updateGroupHeaderSelectionByOptions(this.selectionGroupOptions, this.selectedModel);
     }
     this.updateTriggerText();
   }
@@ -111,43 +102,24 @@ export class SelectComponent extends BaseInputElement implements OnInit {
   }
 
   notifySelectionIds(): void {
-    const value = this.isMultiSelect
-      ? this.selectModelService.getSelectedOptionIds(this.selectedModel as SelectionOption[])
-      : [(this.selectedModel as SelectionOption).id];
+    const value = this.selectModelService.getSelectedOptionIds(this.selectedModel);
     this.selectChange.emit(value);
     this.mySelect.close();
   }
 
-  private onMultiSelectOptionClick(selectedOption: SelectionOption): void {
-    if (selectedOption.isGroupHeader) {
-      this.onGroupHeaderClick(selectedOption);
-    } else {
-      this.selectedModel = this.selectModelService
-        .updateGroupHeaderSelectionByOptions(this.selectionGroupOptions, this.selectedModel as SelectionOption[]);
-    }
-  }
-
   private onGroupHeaderClick(selectedOption: SelectionOption): void {
     const isOptionSelected = this.selectModelService
-      .isOptionSelected(selectedOption, this.selectedModel as SelectionOption[]);
+      .isOptionSelected(selectedOption, this.selectedModel);
 
     this.selectedModel = isOptionSelected
       ? this.selectModelService
-        .selectAllGroupOptions(selectedOption, this.selectionGroupOptions, this.selectedModel as SelectionOption[])
+        .selectAllGroupOptions(selectedOption, this.selectionGroupOptions, this.selectedModel)
       : this.selectModelService
-        .unselectAllGroupOptions(selectedOption, this.selectionGroupOptions, this.selectedModel as SelectionOption[]);
-  }
-
-  private onSingleSelectOptionClick(selectedOption: SelectionOption): void {
-    if (!selectedOption.isGroupHeader) {
-      this.notifySelectionIds();
-    }
+        .unselectAllGroupOptions(selectedOption, this.selectionGroupOptions, this.selectedModel);
   }
 
   private updateTriggerText(): void {
-    this.triggerValue = this.isMultiSelect
-      ? join(flatMap(this.selectedModel, 'value'), ', ')
-      : get(this.selectedModel, 'value', '');
+    this.triggerValue = join(flatMap(this.selectedModel, 'value'), ', ');
     setTimeout(() => {
       this.hasEllipsis = this.triggerValueText.nativeElement.offsetWidth < this.triggerValueText.nativeElement.scrollWidth;
     });
