+++ projects/ui-framework/src/lib/form-elements/select/select.interface.ts
