+++ projects/ui-framework/src/lib/form-elements/select/single-select/single-select.component.html
@@ -1,0 +1,84 @@
+<mat-form-field [ngClass]="{
+    'required': required,
+    'error': errorMessage,
+    'disabled': disabled
+  }">
+
+  <mat-select #mySelect
+              [(ngModel)]="selectedModel"
+              [disabled]="disabled || blockSelectClick"
+              [multiple]="false"
+              [panelClass]="panelClass"
+              disableOptionCentering
+              disableRipple="true">
+
+    <mat-select-trigger
+      fxLayout="row" fxLayoutAlign="start center">
+
+      <b-icon (click)="clearSelection()"
+              (mouseenter)="blockSelectClick=true"
+              (mouseleave)="blockSelectClick=false"
+              [color]="iconColor"
+              [icon]="resetIcon"
+              [size]="iconSize"
+              class="clear-selection">
+      </b-icon>
+
+      <!-- todo: add tooltip for ellipsis -->
+      <span #triggerValueText class="trigger-value">{{triggerValue}}</span>
+
+    </mat-select-trigger>
+
+    <b-search (keydown)="onSearchKeyDown($event)"
+              (searchChange)="onSearchChange($event)"
+              [hideLabelOnFocus]="true"
+              label="Search">
+    </b-search>
+
+    <div [ngClass]="{'no-grouping': selectionGroupOptions.length === 1 && !showSingleGroupHeader}"
+         class="options-wrapper">
+
+      <div *ngFor="let group of selectionGroupOptions"
+           class="group">
+
+        <mat-option (click)="onOptionClick(group.groupHeader)"
+                    *ngIf="selectionGroupOptions.length > 1 || showSingleGroupHeader"
+                    [disabled]="true"
+                    [value]="group.groupHeader"
+                    class="group-header">
+          <span class="option-value">{{group.groupHeader.value}}</span>
+          <div (click)="onCollapseGroup(group)"
+               (mouseenter)="blockGroupHeaderOptionClick=true"
+               (mouseleave)="blockGroupHeaderOptionClick=false"
+               [ngClass]="{'collapse': group.isCollapsed}"
+               class="collapse-group">
+          </div>
+        </mat-option>
+
+        <div [hidden]="group.isCollapsed"
+             class="group-options">
+          <mat-option (click)="onOptionClick(option)"
+                      *ngFor="let option of group.options"
+                      [value]="option"
+                      class="group-option">
+            <span class="option-value">{{option.value}}</span>
+          </mat-option>
+        </div>
+
+      </div>
+
+    </div>
+
+  </mat-select>
+
+  <mat-hint *ngIf="errorMessage" class="error-message">
+    {{errorMessage}}
+  </mat-hint>
+
+  <mat-hint *ngIf="hintMessage && !errorMessage" class="hint-message">
+    {{hintMessage}}
+  </mat-hint>
+
+  <mat-label>{{label}}</mat-label>
+
+</mat-form-field>
