+++ projects/ui-framework/src/lib/form-elements/select/single-select/single-select.component.ts
@@ -1,0 +1,102 @@
+import { Component, EventEmitter, Input, OnInit, Output, ViewChild } from '@angular/core';
+import { concat, escapeRegExp, flatMap, forEach, get, join } from 'lodash';
+import { SelectGroupOption, SelectionGroupOption, SelectionOption } from '../select.interface';
+import { SelectModelService } from '../select-model-service/select-model.service';
+import { BaseInputElement } from '../../base-input-element';
+import { IconColor, Icons, IconSize } from '../../../icons';
+
+const navigationKeys = new Set(['ArrowUp', 'ArrowDown', 'Enter']);
+
+@Component({
+  selector: 'b-single-select',
+  templateUrl: './single-select.component.html',
+  styleUrls: ['../select.scss'],
+})
+export class SingleSelectComponent extends BaseInputElement implements OnInit {
+
+  @ViewChild('mySelect') mySelect;
+  @ViewChild('triggerValueText') triggerValueText;
+
+  @Input() options: SelectGroupOption[];
+  @Input() value: (string | number);
+  @Input() showSingleGroupHeader = false;
+
+  @Output() selectChange: EventEmitter<number | string> = new EventEmitter<number | string>();
+
+  selectionGroupOptions: SelectionGroupOption[];
+  selectedModel: SelectionOption;
+  triggerValue: string;
+  blockGroupHeaderOptionClick = false;
+  blockSelectClick = false;
+  hasEllipsis = false;
+  panelClass: 'single-select-panel';
+  isMultiSelect = false;
+
+  readonly resetIcon: String = Icons.reset_x;
+  readonly iconSize: String = IconSize.small;
+  readonly iconColor: String = IconColor.dark;
+
+  constructor(
+    private selectModelService: SelectModelService,
+  ) {
+    super();
+  }
+
+  ngOnInit(): void {
+    this.selectionGroupOptions = this.selectModelService
+      .getSelectElementOptionsModel(this.options);
+
+    const selectedOptions = this.selectModelService
+      .getSelectedOptions(this.selectionGroupOptions, [this.value]);
+
+    this.selectedModel = selectedOptions[0];
+
+    this.updateTriggerText();
+  }
+
+  onOptionClick(selectedOption): void {
+    if (!selectedOption.isGroupHeader) {
+      this.notifySelectionIds();
+    }
+    this.updateTriggerText();
+  }
+
+  onSearchKeyDown($event): void {
+    if (!navigationKeys.has($event.code)) {
+      $event.stopPropagation();
+    }
+  }
+
+  onCollapseGroup(group): void {
+    group.isCollapsed = !group.isCollapsed;
+  }
+
+  onSearchChange(searchVal: string): void {
+    // todo: filter values by inputEvent.value
+    // const matcher = new RegExp(escapeRegExp(searchVal), 'i');
+    // forEach(this.selectionGroupOptions, (group) => {
+    //   console.log('isMatch', group.groupHeader.value.match(matcher));
+    // });
+  }
+
+  clearSelection(): void {
+    this.selectedModel = null;
+    this.selectChange.emit(null);
+    setTimeout(() => {
+      this.blockSelectClick = false;
+    });
+  }
+
+  notifySelectionIds(): void {
+    const value = (this.selectedModel as SelectionOption).id;
+    this.selectChange.emit(value);
+    this.mySelect.close();
+  }
+
+  private updateTriggerText(): void {
+    this.triggerValue = get(this.selectedModel, 'value', '');
+    setTimeout(() => {
+      this.hasEllipsis = this.triggerValueText.nativeElement.offsetWidth < this.triggerValueText.nativeElement.scrollWidth;
+    });
+  }
+}
