+++ projects/ui-framework/src/lib/overlay/dialog/dialog-example.module.ts
@@ -6,7 +6,7 @@ import { DialogService } from './dialog.service';
 import { MAT_DIALOG_DATA } from '@angular/material';
 import { ButtonsModule } from '../../buttons-indicators/buttons/buttons.module';
 import { FormElementsModule } from '../../form-elements/form-elements.module';
-import { SelectGroupOption } from '../../form-elements/select-elements/select.interface';
+import { SelectGroupOption } from '../../form-elements/select/select.interface';
 import { InputEvent, InputEventType } from '../../form-elements/input';
 
 @Component({
